feat: refactors schema structure, divides typeDefs by concern
