--- apollo/src/schema/index.js
@@ -1,12 +1,0 @@
-import User from './User'
-import schema from './schema'
-
-// console.log('User', User)
-// console.log('typeDefs schema', schema)
-
-const typeDefs = {
-  User,
-  schema
-}
-
-export default typeDefs
