+++ apollo/src/typeDefs/auth/index.js
@@ -1,0 +1,11 @@
+import { inputDef } from './input'
+import { mutationDef } from './mutation'
+import { typeDef } from './type'
+
+
+
+export default {
+  inputDef,
+  mutationDef,
+  typeDef
+}
