+++ apollo/src/typeDefs/auth/input.js
@@ -1,0 +1,12 @@
+import { gql } from 'apollo-server'
+
+
+const inputDef = gql`
+  input LoginInput {
+		email: String!
+		password: String!
+	}
+
+`
+
+export default inputDef
