+++ apollo/src/typeDefs/auth/mutation.js
@@ -1,0 +1,14 @@
+import { gql } from 'apollo-server'
+
+
+const mutationDef = gql`
+  extend type Mutation {
+    login(
+      input: LoginInput!
+    ): AuthPayload!
+  } 
+
+`
+
+
+export default mutationDef
