+++ apollo/src/typeDefs/auth/type.js
@@ -1,0 +1,12 @@
+import { gql } from 'apollo-server'
+
+
+const typeDef = gql`
+  type AuthPayload {
+    token: String!
+    user: User!
+  }
+`
+
+
+export default typeDef
