+++ apollo/src/typeDefs/enum.js
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-server'
+
+const enumDef = gql`
+  enum Microservice {
+    RESUMEQ
+    INERVIEWQ
+  }
+`
+export default enumDef
