+++ apollo/src/typeDefs/job/index.js
@@ -1,0 +1,12 @@
+// import { inputDef } from './input'
+// import { mutationDef } from './mutation'
+import { typeDefs } from './type'
+// import { queryDef } from './query'
+
+
+export default {
+  // inputDef,
+  // mutationDef,
+  typeDefs,
+  // queryDef
+}
