+++ apollo/src/typeDefs/listing/index.js
@@ -1,0 +1,12 @@
+import { inputDefs } from './input'
+import { mutationDefs } from './mutation'
+import { typeDefs } from './type'
+// import { queryDef } from './query'
+
+
+export default {
+  inputDefs,
+  mutationDefs,
+  typeDefs,
+  // queryDef
+}
