+++ apollo/src/typeDefs/query.js
@@ -1,0 +1,12 @@
+import { gql } from 'apollo-server'
+
+
+const query = gql`
+  type Query {
+    info: String!
+  }
+`
+
+export default {
+  query
+}
