+++ apollo/src/typeDefs/scaffold/index.js
@@ -1,0 +1,12 @@
+import { inputDef } from './input'
+import { mutationDef } from './mutation'
+import { typeDef } from './type'
+import { queryDef } from './query'
+
+
+export default {
+  inputDef,
+  mutationDef,
+  typeDef,
+  queryDef
+}
