+++ apollo/src/typeDefs/scaffold/input.js
@@ -1,0 +1,8 @@
+import { gql } from 'apollo-server'
+
+// ! Make plural where multiple mutations are contained
+const inputDef = gql`
+
+`
+
+export default inputDef
