+++ apollo/src/typeDefs/scaffold/mutation.js
@@ -1,0 +1,10 @@
+import { gql } from 'apollo-server'
+
+// ! Make plural where multiple mutations are contained
+const mutationDef = gql`
+  extend type Mutation {
+    
+  }
+`
+
+export default mutationDef
