+++ apollo/src/typeDefs/scaffold/type.js
@@ -1,0 +1,8 @@
+import { gql } from 'apollo-server'
+
+
+// ! Make plural where multiple types are contained
+const typeDef = gql`
+
+`
+export default typeDef
