+++ apollo/src/typeDefs/stripe/query.js
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-server'
+
+const queryDef = gql`
+  extend type Query {
+    stripeBalance: Balance!
+  }
+`
+
+export default queryDef
