+++ apollo/src/typeDefs/stripe/type.js
@@ -1,0 +1,14 @@
+import { gql } from 'apollo-server'
+
+const typeDefs = gql`
+  type Balance {
+		available: Int!
+		pending: Int!
+	}
+
+	type Status {
+		success: String
+		error: String
+	}
+`
+export default typeDefs
